# Speed up remove overlap alignment